# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Technical Test Junior Fullstack (Go + Postgres + UUID)",
        "title": "Inventory Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Fadhail Athaillah Bima Dharmawan",
            "email": "bimadharmawan6@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/marketplace": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Melihat daftar barang yang dijual oleh Seller",
                "tags": [
                    "Marketplace"
                ],
                "summary": "(Pembeli) Lihat Marketplace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search product name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category/Product Type ID",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum price",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum price",
                        "name": "max_price",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Melihat semua master produk (Admin \u0026 Seller bisa lihat)",
                "tags": [
                    "Product Master (Gudang)"
                ],
                "summary": "Lihat Daftar Barang Gudang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search product name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Product Type ID filter",
                        "name": "product_type_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin memasukkan master data produk",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Master (Gudang)"
                ],
                "summary": "Input Barang ke Gudang (Admin)",
                "parameters": [
                    {
                        "description": "Data Produk",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateProductInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/low-stock": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get products with stock below threshold",
                "tags": [
                    "Product Master (Gudang)"
                ],
                "summary": "Get Low Stock Products (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stock threshold (default: 10)",
                        "name": "threshold",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin dapat memperbarui data produk master (nama, stock, harga, kategori)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Master (Gudang)"
                ],
                "summary": "Update Barang Gudang (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateProductInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Product Master (Gudang)"
                ],
                "summary": "Hapus Barang Gudang (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/seller/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Melihat daftar produk yang dijual oleh seller yang sedang login",
                "tags": [
                    "Seller Catalog"
                ],
                "summary": "(Seller) Lihat Daftar Produk Sendiri",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Seller memilih barang dari gudang admin dan menentukan harga jual sendiri",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seller Catalog"
                ],
                "summary": "(Seller) Pajang Barang \u0026 Markup Harga",
                "parameters": [
                    {
                        "description": "Data Markup",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.AddToEtalaseInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/seller/products/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Seller dapat memperbarui harga jual atau status aktif produk mereka",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seller Catalog"
                ],
                "summary": "(Seller) Update Harga Produk di Marketplace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Seller Product ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateSellerProductInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Seller menonaktifkan produk mereka dari marketplace (soft delete)",
                "tags": [
                    "Seller Catalog"
                ],
                "summary": "(Seller) Hapus Produk dari Marketplace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Seller Product ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "services.AddToEtalaseInput": {
            "type": "object",
            "required": [
                "product_id",
                "selling_price"
            ],
            "properties": {
                "product_id": {
                    "description": "UUID produk dari gudang pusat",
                    "type": "string"
                },
                "selling_price": {
                    "description": "Harga jual seller",
                    "type": "number",
                    "minimum": 1
                }
            }
        },
...
        "services.CreateProductInput": {
            "type": "object",
            "required": [
                "name",
                "price",
                "product_type_id",
                "stock"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "minimum": 1
                },
                "product_type_id": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "services.UpdateProductInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "minimum": 1
                },
                "product_type_id": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "services.UpdateSellerProductInput": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "selling_price": {
                    "type": "number",
                    "minimum": 1
                }
            }
        },
...
    }
}